Add last resort section to Troubleshooting
As a last resort, an Arch user can configure pacman to
ignore the GPG verfication step. This is quite dangerous so
once the system is fully upgraded the user should change it
back. This, of course, should be only required if none of
the previous attempts to upgrade the system worked. Very
weird how Arch maintainers simply refuse to fix this crap.

Signed-off-by: Lincoln Yuji <lincolnyuji@hotmail.com>