_posts: Update disclaimer at tmpfs cache blog post
The issue was partially solved. So I believe it's good to
let that regisred in the blogpost. However, it's important to
note that the stuttering seems to be still an issue. What was
solved was the desktop environment completely hanging whenever
an animation loads from a slow or busy channel from drives.

Signed-off-by: Lincoln Yuji <lincolnyuji@hotmail.com>